LivingAtlas1-main/backend/requirements.txt
aiofiles==23.2.1
alembic==1.10.3
anyio==3.6.2
apyori==1.1.2
async-generator==1.10
attrs==21.4.0
cachetools==5.3.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.3
cryptography==36.0.1
exceptiongroup==1.2.0
fastapi==0.95.0
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
h11==0.13.0
httpcore==1.0.2
httpx==0.25.2
idna==3.3
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.23.3
outcome==1.1.0
packaging==23.2
pluggy==1.3.0
protobuf==4.25.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pyOpenSSL==22.0.0
pyperclip==1.8.2
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.2.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.8
starlette==0.26.1
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.8
uvicorn==0.21.1
wsproto==1.0.0
